LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/manual-scroll-reload-expected.txt


FAIL scroll: scroll() should work on a reload navigation assert_equals: expected 18 but got 431
